refactor(ner): align the candle backend with house conventions
Errors now use XbergError::Plugin with plugin_name ner-candle plus
validation, matching gline.rs, instead of blanket XbergError::Other.
Model load and cache hits emit the same tracing events the ONNX backend
emits. The seven v2_*.rs files move into a v2/ module whose doc states
what v2 means: the GLiNER2 schema-prompt model architecture, not a
revision of this crate. Public API is unchanged; re-exports keep
Gliner2, V2Tokenizer, V2Splitter, encode_v2 where they were.
